requirements.txt
streamlit
groq
python-dotenv
pymupdf
networkx
matplotlib